requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
cassandra-driver==3.29.2
weaviate-client==4.9.5
google-generativeai==0.8.3
python-dotenv==1.0.1
pyyaml==6.0.2
